fix(wxo): keep mapper imports free of IBM SDK via payloads Move Watsonx naming and field validation helpers into payloads.py so mapper -> payloads no longer imports utils.py and its IBM client deps. Point core modules and tests at the shared payload helpers and leave utils.py for runtime error/utility helpers only.
